Genericize Biome db interactions
This allows for database interactions in the Biome module to be db
backend agnostic.
